pyproject.toml
[project]
name = "Psihosomatica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.25.0",
    "alembic>=1.18.4",
    "alembic-postgresql-enum>=1.10.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "celery[redis]>=5.3.1",
    "cryptography>=46.0.5",
    "emails>=0.6",
    "fastapi>=0.133.0",
    "imgkit>=1.2.3",
    "paginate-sqlalchemy>=0.3.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyfcm>=2.1.0",
    "pymorphy2>=0.9.1",
    "pymorphy2-dicts>=2.4.393442.3710985",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "pytz>=2025.2",
    "redis>=7.2.0",
    "rtree>=1.4.1",
    "setuptools>=82.0.0",
    "sqlalchemy[mypy]>=2.0.46",
    "tenacity>=9.1.4",
    "user-agents>=2.2.0",
    "uvicorn>=0.41.0",
]
